requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
distlib==0.3.8
Django==4.2.10
django-daraja==1.3.0
django-dotenv==1.4.2
django-extensions==3.2.3
et-xmlfile==1.1.0
filelock==3.13.1
graphviz==0.20.1
idna==3.6
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pycparser==2.21
pydotplus==2.0.2
pygraphviz==1.11
pyparsing==3.1.1
pyproject-api==1.6.1
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
stripe==8.5.0
tomli==2.0.1
tox==4.12.1
typing-extensions==4.9.0
urllib3==2.2.1
virtualenv==20.25.0
